🐛 fix(curveframes): promote the curve parameter at `_param`, preserving f32 (#899)
Four accessors promoted with `astype(float)`, which names the *default* float
-- so under `JAX_ENABLE_X64=1` an f32 `tau` or `station` was silently widened
to f64 and a deliberate choice of single precision discarded.

All four did the same two lines right after `_param`, and `_param` is the one
funnel they route through -- as its own docstring already said. So the
promotion moved there and the casts went away: a net deletion. `base.tangent`
was casting and then calling `rotation_matrix`, which cast again; that
double-cast is gone rather than replaced, and `signedplanar._float_param` is
dropped in favour of the shared helper.

`BishopBuilder._tangent_at` keeps an explicit call: three of its five callers
build their own `Quantity` (the ODE seed, the integration variable, the vmap)
rather than resolving one through `_param`.

Bishop's `rotation_matrix` stays f64, and that is correct rather than an
unfixed gap: the transport ODE runs at a default `rtol = atol = 1e-10`, three
orders of magnitude below f32's eps of ~1.2e-7, so the solve could not meet
its own tolerance in single precision. Pinned as a test asserting f64 with
that reasoning, so a later "fix" trips a guard instead of quietly breaking the
tolerance contract.

Closes #886.

+++ packages/coordinaxs.curveframes/src/coordinaxs/curveframes/_src/frenetserret.py
@@ -187,7 +187,6 @@ class FrenetSerretBuilder(AbstractCurveFrameBuilder):
         # parameter does not have, so the other order reported the accident
         # rather than the cause.
         g, tau_unit = b._param(p)
-        g = g.astype(float)
         dcurve = u.experimental.jacfwd(b.curve, units=(tau_unit,))
         d2curve = u.experimental.jacfwd(dcurve, units=(tau_unit,))
 
@@ -245,7 +244,7 @@ class FrenetSerretBuilder(AbstractCurveFrameBuilder):
         b, p = self._resolve(tau)
         g, tau_unit = b._param(p)
         dcurve = u.experimental.jacfwd(b.curve, units=(tau_unit,))
-        return u.Q(_normalize(dcurve(g.astype(float))).value, "")
+        return u.Q(_normalize(dcurve(g)).value, "")
 
     def normal(self, tau: Any, /) -> u.Q:
         r"""Return the unit normal vector $\mathbf{N}(\tau)$ (row 1 of R).
